requirements-Eylons-Laptop.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.28.0
openai>=1.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
